requirements.txt
fastapi[all]
sqlmodel
mysqlclient

# Use for sending mail
fastapi-mail

# Use for create username
Unidecode

# Use for Token JTW
python-jose[cryptography]

# Use for password
passlib[bcrypt]